in xscatter3d.c: for VENDOR BLAS, add string length "1" at the end of GEMM calls. in pxgstrf3d.c:    initialize stat->peak_buffer = 0;    remove xinitDiagFactBufs(ldt, &dFBuf); dFBuf is not used.    remove initCommRequests(&comReqs, grid); comReqs is not used.    remove initMsgs(&msgs); msgs is not used.    add calls freePackLUInfo(), etc. to deallocate memory. in superlu_grid3d.c: add MPI_Comm_free(&superlu3d_comm); in px3dcomm.c:: xGatherAllFactoredLU():    add SUPERLU_FREE(gNodeCount); SUPERLU_FREE(gNodeLists); in xtrfAux.c:    xinit3DLUstructForest(): add SUPERLU_FREE(gNodeCount); SUPERLU_FREE(gNodeLists);    xinitTrf3Dpartition(): add SUPERLU_FREE(perm_c_supno); in treeFactorization.c: add functions:    freePackLUInfo(packLUInfo_t* packLUInfo);    freeFactStat(factStat_t* factStat);    freeFactNodelists(factNodelists_t *fNlists);    freeMsgsArr(int_t numLA, msgs_t **msgss);    freeCommRequestsArr(); in xtreeFactorization.c: add functions:    xLluBufFreeArr();    xfreeDiagFactBufsArr(mxLeafNode, dFBufs);    xfreeScuBufs(scuBufs_t* scuBufs); in supernodalForest.c :: calcTopInfoForest():    add SUPERLU_FREE(myTopOrder); // sherry added in pxgssvx3d.c: add call to SCT_free(); in sec_structs.h: add free(line);

+++ SRC/dreadtriple_noheader.c
@@ -87,7 +87,7 @@ dreadtriple_noheader(FILE *fp, int_t *m, int_t *n, int_t *nonz,
 #endif
 
     /* Second pass: read the actual matrix values */
-    printf("m %ld, n %ld, nonz %ld\n", *m, *n, *nonz);
+    printf("m %ld, n %ld, nonz %ld\n", (long int) *m, (long int) *n, (long int) *nonz);
     dallocateA_dist(*n, new_nonz, nzval, rowind, colptr); /* Allocate storage */
     a    = *nzval;
     asub = *rowind;
